Add constructor using specifed remote address
